+++ src/RGen.Domain/GeneratorService.cs
@@ -20,8 +20,8 @@ public class GeneratorService : IGeneratorService
 		int numberOfElements,
 		int numberOfSets,
 		int? length,
-		int? min,
-		int? max,
+		ulong? min,
+		ulong? max,
 		CancellationToken cancellationToken = default)
 	{
 		var sets = generator.Generate(numberOfElements, numberOfSets, length, min, max);
